refactor(core): share the remote thread list empty state and seeding
