+++ .changeset/remote-thread-seed.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/core": patch
+---
+
+refactor: share the remote thread list empty state and local thread seeding.
